+++ src/BLAS_LAPACK/R_BLAS_LAPACK.cpp
@@ -1,19 +1,16 @@
 #include <Rcpp.h>
-#include <R_ext/BLAS.h>
+#include "../Rconfig-wrap.h"
 #include <R_ext/Lapack.h>
 #include "../R_BLAS_LAPACK.h"
+
+static const char C_N = 'N', C_L = 'L';
 
 namespace R_BLAS_LAPACK {
   extern "C"
   {
-    // Non LAPACK function to make rank one update of of chol decomp
-    // We could use the LINPACK function dchex. See
-    //  http://www.netlib.org/linpack/dchex.f
-    // See http://icl.cs.utk.edu/lapack-forum/viewtopic.php?f=2&t=2646
-    // I use the macro from r-source/src/include/R_ext/RS.h
     void F77_NAME(dchur)(
-        int*,   // UPPER
-        int*,   // DOTRAN
+        const char*,   // UPLO
+        const char*,   // TRANS
         int*,    // N
         int*,    // M
         double*, // R
@@ -25,8 +22,14 @@ namespace R_BLAS_LAPACK {
         double*, // RHO
         double*, // C
         double*, // S
-        int*     // INFO
+        int*     // INFO,
+        FCLEN FCLEN
     );
+
+    void F77_NAME(dsyr)(
+      const char *uplo, const int *n, const double *alpha,
+      const double *x, const int *incx,
+      double *a, const int *lda FCLEN);
   }
 
   void ddhazard_dchur(double *R, double *x, int n, int ldr){
@@ -41,14 +44,11 @@ namespace R_BLAS_LAPACK {
     int ldz = 1;
     double z, y, rho;
 
-    int UPPER = false;
-    int DOTRAN = false;
-
     F77_CALL(dchur)(
-        &UPPER,   // lower triangular
-        &DOTRAN,   // does not matter. Relates to Z
+        &C_L,   // lower triangular
+        &C_N,   // does not matter. Relates to Z
         &n, &m, R, &ldr,
-        x, &z, &ldz, &y, &rho, c, s, &info);
+        x, &z, &ldz, &y, &rho, c, s, &info FCONE FCONE);
 
     delete[] c;
     delete[] s;
@@ -64,7 +64,7 @@ namespace R_BLAS_LAPACK {
     int info;
 
     char c1 = 'L', c2 = 'N';
-    F77_CALL(dtrtri)(&c1, &c2, &n, out, &ldr, &info);
+    F77_CALL(dtrtri)(&c1, &c2, &n, out, &ldr, &info FCONE FCONE);
 
     if(info != 0){
       std::stringstream str;
@@ -77,7 +77,7 @@ namespace R_BLAS_LAPACK {
     int info;
 
     char c1 = 'L';
-    F77_CALL(dpotrf)(&c1, &n, out, &lda, &info);
+    F77_CALL(dpotrf)(&c1, &n, out, &lda, &info FCONE);
 
     if(info != 0){
       std::stringstream str;
@@ -96,7 +96,7 @@ namespace R_BLAS_LAPACK {
     int incx = 1;
     char c1 = 'L', c2 = is_transpose ? 'T' : 'N', c3 = 'N';
     F77_CALL(dtrmv)(
-        &c1, &c2, &c3, &n, A, &lda, x, &incx);
+        &c1, &c2, &c3, &n, A, &lda, x, &incx FCONE FCONE FCONE);
   }
 
   void sym_mat_rank_one_update(
@@ -117,7 +117,7 @@ namespace R_BLAS_LAPACK {
     F77_CALL(dsyr)(
         &c1, n, alpha,
         x, &inx,
-        A, n);
+        A, n FCONE);
   }
 
   void sym_mat_vec_mult(
@@ -143,7 +143,7 @@ namespace R_BLAS_LAPACK {
         &c1, n,
         &dum_d, A, n,
         x, &dum_i, &dum_d,
-        y, &dum_i);
+        y, &dum_i FCONE);
   }
 
   void symmetric_rank_k_update(
@@ -172,7 +172,7 @@ namespace R_BLAS_LAPACK {
         &c1, &c2,
         n, k, &dum_d,
         A, n,
-        &dum_d, C, n);
+        &dum_d, C, n FCONE FCONE);
   }
 
   void triangular_sys_solve(
@@ -195,7 +195,7 @@ namespace R_BLAS_LAPACK {
         &n, &nrhs,
         A, &n,
         B, &n,
-        &info);
+        &info FCONE FCONE FCONE);
 
     if(info != 0){
       std::stringstream str;
@@ -211,7 +211,7 @@ namespace R_BLAS_LAPACK {
       double *b, const int *ldb){
     F77_CALL(dtrmm)(
         side, uplo, transa, diag, m, n, alpha, a, lda,
-        b, ldb);
+        b, ldb FCONE FCONE FCONE FCONE);
   }
 
   void dtrmv(
@@ -219,7 +219,7 @@ namespace R_BLAS_LAPACK {
       const int *n, const double *a, const int *lda,
       double *x, const int *incx){
     F77_CALL(dtrmv)(
-        uplo, trans, diag, n, a, lda, x, incx);
+        uplo, trans, diag, n, a, lda, x, incx FCONE FCONE FCONE);
   }
 
   void dger(
@@ -249,7 +249,7 @@ namespace R_BLAS_LAPACK {
   void dgetrs(
       const char* trans, const int* n, const int* nrhs, const double* a,
       const int* lda, const int* ipiv, double* b, const int* ldb, int* info){
-    F77_CALL(dgetrs)(trans, n, nrhs, a, lda, ipiv, b, ldb, info);
+    F77_CALL(dgetrs)(trans, n, nrhs, a, lda, ipiv, b, ldb, info FCONE);
   }
 
   void dormqr(const char* side, const char* trans,
@@ -258,7 +258,8 @@ namespace R_BLAS_LAPACK {
               const double* tau, double* c, const int* ldc,
               double* work, const int* lwork, int* info){
     F77_CALL(dormqr)(
-        side, trans, m, n, k, a, lda, tau, c, ldc, work, lwork, info);
+        side, trans, m, n, k, a, lda, tau, c, ldc, work, lwork, info
+        FCONE FCONE);
   }
 
   void dgeqp3(const int* m, const int* n, double* a, const int* lda,
@@ -270,5 +271,11 @@ namespace R_BLAS_LAPACK {
   void dgetri(const int* n, double* a, const int* lda,
               int* ipiv, double* work, const int* lwork, int* info){
     F77_CALL(dgetri)(n, a, lda, ipiv, work, lwork, info);
+  }
+
+  void dsyr(const char *uplo, const int *n, const double *alpha,
+            const double *x, const int *incx,
+            double *a, const int *lda){
+    F77_CALL(dsyr)(uplo, n, alpha, x, incx, a, lda FCONE);
   }
 }